Extract some utils
